goku-terminal-agent/requirements.txt
fastapi
uvicorn
python-dotenv
litellm
requests
pydantic
sqlalchemy
psycopg2-binary
redis
celery
qdrant-client
mcp
python-multipart
python-jose[cryptography]
passlib[bcrypt]
textual
rich
typer
websockets
httpx
